pyproject.toml
[tool.poetry]
name = "javelin-sdk"
version = "RELEASE_VERSION"
description = "Python client for Highflame"
authors = ["Sharath Rajasekar <sharath@highflame.com>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://getjavelin.com"
packages = [
    { include = "javelin_cli" },
    { include = "javelin_sdk" },
]

[tool.poetry.scripts]
javelin = "javelin_cli.cli:main"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.2"
pydantic = "^2.9.2"
requests = "^2.32.3"
urllib3 = ">=2.2.2,<3.0.0"
jmespath = "^1.0.1"
jsonpath-ng = "^1.7.0"

# OpenTelemetry Dependencies
opentelemetry-api = "^1.32.1"
opentelemetry-sdk = "^1.32.1"
opentelemetry-semantic-conventions = "^0.53b1"
opentelemetry-exporter-otlp-proto-http = "^1.32.1"
opentelemetry-exporter-otlp-proto-grpc = "^1.32.1"

# opentelemetry-instrumentation = "^0.53b1"
# opentelemetry-instrumentation-botocore = "^0.53b1"
# opentelemetry-util-http = "^0.53b1"
# opentelemetry-instrumentation-asgi = "^0.53b1"
# opentelemetry-instrumentation-fastapi = "^0.53b1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
pytest-httpx = "^0.32.0"
pytest-asyncio = "^0.21.0"
pytest-mock = "^3.10.0"

[tool.poetry.group.dev.dependencies]
black = "24.3.0"
flake8 = "^7.3.0"
pre-commit = "^3.3.1"
mkdocs = "^1.4.3"
mkdocstrings = {version = "0.21.2", extras = ["python"]}
python-dotenv = "^1.0.0"
mkdocs-material = "^9.6.11"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
